Added In UI, space craft position options, and time constraint options
Added in UI to make different decisions of how the simulation will run. Created a sign variable to indicate if the spacecraft is infront of or behind the direction of the asteroid motion. Also added a time constraint so the simulation can be run without using the thrusters at all and only run with the spacecraft near the asteroid.