Massive Change Update
Added deflection in the B-Plane, trying to validate against PDC paper but conditions are slightly off due to thruster turning off due to power in the paper and that has not been implemented in this code yet. Also fixed a minor bug with the placement of the sc_num scaling thrust.